requirements.txt
psycopg2~=2.9.9
pandas = 1.5.1
sqlalchemy =  2.0.26

pandas~=2.2.2
python-dateutil~=2.9.0.post0